scm: split http-api and database tests
